+++ service/handlers.go
@@ -3,12 +3,14 @@ package main
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 	"strings"
 
 	"github.com/cello-proj/cello/internal/requests"
+	"github.com/cello-proj/cello/internal/responses"
 	"github.com/cello-proj/cello/internal/types"
 	"github.com/cello-proj/cello/service/internal/credentials"
 	"github.com/cello-proj/cello/service/internal/db"
@@ -19,7 +21,12 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/gorilla/mux"
+	upper "github.com/upper/db/v4"
 	"gopkg.in/yaml.v2"
+)
+
+const (
+	numOfTokensLimit = 2
 )
 
 // Represents a JWT token.
@@ -489,6 +496,39 @@ func newCelloToken(provider, key, secret string) *token {
 	}
 }
 
+// projectExists checks if a project exists using both the credential provider and database
+func (h handler) projectExists(ctx context.Context, l log.Logger, cp credentials.Provider, w http.ResponseWriter, projectName string) (bool, error) {
+
+	// Checking credential provider
+	level.Debug(l).Log("message", "checking if project exists")
+	projectExists, err := cp.ProjectExists(projectName)
+	if err != nil {
+		level.Error(l).Log("message", "error checking credentials provider for project", "error", err)
+		h.errorResponse(w, "error retrieving project", http.StatusInternalServerError)
+		return false, err
+	}
+
+	if !projectExists {
+		level.Debug(l).Log("message", "project does not exist in credentials provider")
+		h.errorResponse(w, "project does not exist", http.StatusNotFound)
+		return false, err
+	}
+
+	// Checking database
+	_, err = h.dbClient.ReadProjectEntry(ctx, projectName)
+	if err != nil {
+		level.Error(l).Log("message", "error retrieving project from database", "error", err)
+		if errors.Is(err, upper.ErrNoMoreRows) {
+			h.errorResponse(w, "project does not exist", http.StatusNotFound)
+		} else {
+			h.errorResponse(w, "error retrieving project", http.StatusInternalServerError)
+		}
+		return false, err
+	}
+
+	return true, err
+}
+
 // Creates a project
 func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 	l := h.requestLogger(r, "op", "create-project")
@@ -559,22 +599,34 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	level.Debug(l).Log("message", "creating project")
-	role, secret, err := cp.CreateProject(capp.Name)
+	role, secret, secretAccessor, err := cp.CreateProject(capp.Name)
 	if err != nil {
 		level.Error(l).Log("message", "error creating project", "error", err)
 		h.errorResponse(w, "error creating project", http.StatusInternalServerError)
 		return
 	}
 
-	level.Debug(l).Log("message", "retrieving Cello token")
-	t := newCelloToken("vault", role, secret)
-	jsonResult, err := json.Marshal(t)
+	level.Debug(l).Log("message", "inserting token into DB")
+	_, err = h.dbClient.CreateTokenEntry(ctx, capp.Name, secretAccessor)
 	if err != nil {
+		level.Error(l).Log("message", "error inserting token into DB", "error", err)
+		h.errorResponse(w, "error creating token", http.StatusInternalServerError)
+		return
+	}
+
+	level.Debug(l).Log("message", "retrieving Cello token")
+	token := newCelloToken("vault", role, secret)
+
+	resp := responses.CreateProject{
+		Token:   token.Token,
+		TokenID: secretAccessor,
+	}
+
+	if err := json.NewEncoder(w).Encode(resp); err != nil {
 		level.Error(l).Log("message", "error serializing token", "error", err)
 		h.errorResponse(w, "error serializing token", http.StatusInternalServerError)
 		return
 	}
-	fmt.Fprint(w, string(jsonResult))
 }
 
 // Get a project
@@ -596,30 +648,29 @@ func (h handler) getProject(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	level.Debug(l).Log("message", "creating credential provider")
-	cp, err := h.newCredentialsProvider(*a, h.env, r.Header, credentials.NewVaultConfig, credentials.NewVaultSvc)
-	if err != nil {
-		level.Error(l).Log("message", "error creating credentials provider", "error", err)
-		h.errorResponse(w, "error creating credentials provider", http.StatusInternalServerError)
-		return
-	}
-
-	level.Debug(l).Log("message", "getting project")
-	resp, err := cp.GetProject(projectName)
+	level.Debug(l).Log("message", "getting project from database")
+	ctx := r.Context()
+	projectEntry, err := h.dbClient.ReadProjectEntry(ctx, projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error retrieving project", "error", err)
-		h.errorResponse(w, "error retrieving project", http.StatusNotFound)
+		if errors.Is(err, upper.ErrNoMoreRows) {
+			h.errorResponse(w, "error retrieving project", http.StatusNotFound)
+		} else {
+			h.errorResponse(w, "error retrieving project", http.StatusInternalServerError)
+		}
 		return
 	}
 
-	data, err := json.Marshal(resp)
-	if err != nil {
+	resp := responses.GetProject{
+		Name:       projectName,
+		Repository: projectEntry.Repository,
+	}
+
+	if err := json.NewEncoder(w).Encode(resp); err != nil {
 		level.Error(l).Log("message", "error creating response", "error", err)
 		h.errorResponse(w, "error creating response object", http.StatusInternalServerError)
 		return
 	}
-
-	fmt.Fprint(w, string(data))
 }
 
 // Delete a project
@@ -774,7 +825,6 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 		h.errorResponse(w, "error creating target", http.StatusInternalServerError)
 		return
 	}
-
 	fmt.Fprint(w, "{}")
 }
 
@@ -974,6 +1024,218 @@ func (h handler) updateTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	fmt.Fprint(w, string(data))
+}
+
+func (h handler) deleteToken(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	projectName := vars["projectName"]
+	tokenID := vars["tokenID"]
+
+	l := h.requestLogger(r, "op", "delete-token", "project", projectName, "tokenID", tokenID)
+
+	level.Debug(l).Log("message", "validating authorization header for delete token")
+
+	ah := r.Header.Get("Authorization")
+	a, err := credentials.NewAuthorization(ah)
+	if err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header format", http.StatusUnauthorized)
+		return
+	}
+	if err := a.Validate(a.ValidateAuthorizedAdmin(h.env.AdminSecret)); err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header", http.StatusUnauthorized)
+		return
+	}
+
+	level.Debug(l).Log("message", "creating credential provider")
+	cp, err := h.newCredentialsProvider(*a, h.env, r.Header, credentials.NewVaultConfig, credentials.NewVaultSvc)
+	if err != nil {
+		level.Error(l).Log("message", "error creating credentials provider", "error", err)
+		h.errorResponse(w, "error creating credentials provider", http.StatusInternalServerError)
+		return
+	}
+
+	ctx := r.Context()
+
+	level.Debug(l).Log("message", "checking if project exists")
+	projectExists, err := h.projectExists(ctx, l, cp, w, projectName)
+
+	if err != nil || !projectExists {
+		return
+	}
+
+	// check if token exists in CP and DB
+	projectToken, err := cp.GetProjectToken(projectName, tokenID)
+	if err != nil {
+		level.Error(l).Log("message", "error retrieving token from credentials provider", "error", err)
+		h.errorResponse(w, "error retrieving token", http.StatusInternalServerError)
+		return
+	}
+
+	if projectToken == (types.ProjectToken{}) {
+		level.Error(l).Log("message", "token does not exist in credential provider", "error", err)
+		h.errorResponse(w, "token does not exist", http.StatusNotFound)
+		return
+	}
+
+	if _, err = h.dbClient.ReadTokenEntry(ctx, tokenID); err != nil {
+		if errors.Is(err, upper.ErrNoMoreRows) {
+			level.Error(l).Log("message", "token does not exist in DB", "error", err)
+			h.errorResponse(w, "token does not exist", http.StatusNotFound)
+			return
+		}
+		level.Error(l).Log("message", "error retrieving token from DB", "error", err)
+		h.errorResponse(w, "error retrieving token", http.StatusInternalServerError)
+		return
+	}
+
+	// delete token from DB and CP
+	level.Debug(l).Log("message", "deleting token from database")
+	if err = h.dbClient.DeleteTokenEntry(ctx, tokenID); err != nil {
+		level.Error(l).Log("message", "error deleting token from database", "error", err)
+		h.errorResponse(w, "error deleting token", http.StatusInternalServerError)
+		return
+	}
+
+	level.Debug(l).Log("message", "deleting token from credentials provider")
+	if err = cp.DeleteProjectToken(projectName, tokenID); err != nil {
+		level.Error(l).Log("message", "error deleting token from credentials provider", "error", err)
+		h.errorResponse(w, "error deleting token", http.StatusInternalServerError)
+	}
+}
+
+// Creates a token
+func (h handler) createToken(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	projectName := vars["projectName"]
+
+	l := h.requestLogger(r, "op", "create-token", "project", projectName)
+
+	level.Debug(l).Log("message", "validating authorization header for token create")
+	ah := r.Header.Get("Authorization")
+	a, err := credentials.NewAuthorization(ah)
+	if err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header format", http.StatusUnauthorized)
+		return
+	}
+	if err := a.Validate(a.ValidateAuthorizedAdmin(h.env.AdminSecret)); err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header", http.StatusUnauthorized)
+		return
+	}
+
+	ctx := r.Context()
+
+	level.Debug(l).Log("message", "creating credential provider")
+	cp, err := h.newCredentialsProvider(*a, h.env, r.Header, credentials.NewVaultConfig, credentials.NewVaultSvc)
+	if err != nil {
+		level.Error(l).Log("message", "error creating credentials provider", "error", err)
+		h.errorResponse(w, "error creating credentials provider", http.StatusInternalServerError)
+		return
+	}
+	projectExists, err := h.projectExists(ctx, l, cp, w, projectName)
+
+	if err != nil || !projectExists {
+		return
+	}
+
+	tokens, err := h.dbClient.ListTokenEntries(ctx, projectName)
+	if err != nil {
+		level.Error(l).Log("message", "error listing tokens from DB", "error", err)
+		h.errorResponse(w, "error listing tokens", http.StatusInternalServerError)
+		return
+	}
+
+	if len(tokens) >= numOfTokensLimit {
+		level.Error(l).Log("message", "number of tokens allowed per project has been reached")
+		h.errorResponse(w, "token limit reached", http.StatusInternalServerError)
+		return
+	}
+
+	level.Debug(l).Log("message", "creating token")
+	role, secret, secretAccessor, err := cp.CreateToken(projectName)
+	if err != nil {
+		level.Error(l).Log("message", "error creating token with credentials provider", "error", err)
+		h.errorResponse(w, "error creating token with credentials provider", http.StatusInternalServerError)
+		return
+	}
+
+	level.Debug(l).Log("message", "inserting into db")
+	te, err := h.dbClient.CreateTokenEntry(ctx, projectName, secretAccessor)
+	if err != nil {
+		level.Error(l).Log("message", "error inserting token to db", "error", err)
+		h.errorResponse(w, "error creating token", http.StatusInternalServerError)
+		return
+	}
+
+	token := newCelloToken("vault", role, secret)
+
+	resp := responses.CreateToken{
+		CreatedAt: te.CreatedAt,
+		Token:     token.Token,
+		TokenID:   te.TokenID,
+	}
+
+	if err := json.NewEncoder(w).Encode(resp); err != nil {
+		level.Error(l).Log("message", "error serializing project token", "error", err)
+		h.errorResponse(w, "error listing project tokens", http.StatusInternalServerError)
+		return
+	}
+}
+
+// Lists tokens for a project
+func (h handler) listTokens(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	projectName := vars["projectName"]
+
+	l := h.requestLogger(r, "op", "list-tokens", "project", projectName)
+
+	level.Debug(l).Log("message", "validating authorization header for token list")
+	ah := r.Header.Get("Authorization")
+	a, err := credentials.NewAuthorization(ah)
+	if err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header format", http.StatusUnauthorized)
+		return
+	}
+	if err := a.Validate(a.ValidateAuthorizedAdmin(h.env.AdminSecret)); err != nil {
+		h.errorResponse(w, "error unauthorized, invalid authorization header", http.StatusUnauthorized)
+		return
+	}
+
+	ctx := r.Context()
+
+	level.Debug(l).Log("message", "creating credential provider")
+	cp, err := h.newCredentialsProvider(*a, h.env, r.Header, credentials.NewVaultConfig, credentials.NewVaultSvc)
+	if err != nil {
+		level.Error(l).Log("message", "error creating credentials provider", "error", err)
+		h.errorResponse(w, "error creating credentials provider", http.StatusInternalServerError)
+		return
+	}
+
+	projectExists, err := h.projectExists(ctx, l, cp, w, projectName)
+
+	if err != nil || !projectExists {
+		return
+	}
+
+	tokens, err := h.dbClient.ListTokenEntries(ctx, projectName)
+	if err != nil {
+		level.Error(l).Log("message", "error listing project tokens", "error", err)
+		h.errorResponse(w, "error listing project tokens", http.StatusInternalServerError)
+		return
+	}
+
+	resp := []responses.ListTokens{}
+	for _, tokenEntry := range tokens {
+		resp = append(resp, responses.ListTokens{
+			CreatedAt: tokenEntry.CreatedAt,
+			TokenID:   tokenEntry.TokenID,
+		})
+	}
+
+	if err := json.NewEncoder(w).Encode(resp); err != nil {
+		level.Error(l).Log("message", "error serializing project tokens", "error", err)
+		h.errorResponse(w, "error listing project tokens", http.StatusInternalServerError)
+		return
+	}
 }
 
 // Convenience method that writes a failure response in a standard manner