+++ service/handlers_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -20,17 +21,20 @@ import (
 	"github.com/cello-proj/cello/service/internal/env"
 	"github.com/cello-proj/cello/service/internal/git"
 	"github.com/cello-proj/cello/service/internal/workflow"
+	th "github.com/cello-proj/cello/service/test/testhelpers"
 
 	"github.com/go-kit/log"
 	"github.com/stretchr/testify/assert"
+	upper "github.com/upper/db/v4"
 )
 
 const (
 	// #nosec
-	testPassword      = "D34DB33FD34DB33FD34DB33FD34DB33F"
-	userAuthHeader    = "vault:user:" + testPassword
-	invalidAuthHeader = "bad auth header"
-	adminAuthHeader   = "vault:admin:" + testPassword
+	testPassword        = "D34DB33FD34DB33FD34DB33FD34DB33F"
+	userAuthHeader      = "vault:user:" + testPassword
+	invalidAuthHeader   = "bad auth header"
+	adminAuthHeader     = "vault:admin:" + testPassword
+	projectDoesNotExist = "projectdoesnotexist"
 )
 
 type mockDB struct{}
@@ -47,7 +51,48 @@ func (d mockDB) CreateProjectEntry(ctx context.Context, pe db.ProjectEntry) erro
 	return nil
 }
 
+func (d mockDB) CreateTokenEntry(ctx context.Context, project string, secretAccessor string) (db.TokenEntry, error) {
+	if project == "tokendberror" || project == "tokendbentryerror" {
+		return db.TokenEntry{}, fmt.Errorf("token db error")
+	}
+
+	token := db.TokenEntry{
+		CreatedAt: "2022-06-21T14:56:10.341066-07:00",
+		ProjectID: project,
+		TokenID:   secretAccessor,
+	}
+	return token, nil
+}
+
+func (d mockDB) ListTokenEntries(ctx context.Context, project string) ([]db.TokenEntry, error) {
+	if project == projectDoesNotExist {
+		return []db.TokenEntry{}, upper.ErrNoMoreRows
+	}
+
+	if project == "projectreaderror" {
+		return []db.TokenEntry{}, errors.New("error reading DB")
+	}
+
+	if project == "projectlisttokenserror" {
+		return []db.TokenEntry{}, errors.New("error reading DB")
+	}
+
+	if project == "projectlisttokenslimit" {
+		return []db.TokenEntry{{ProjectID: "project1", TokenID: "1234"}, {ProjectID: "project1", TokenID: "5678"}}, nil
+	}
+
+	return []db.TokenEntry{}, nil
+}
+
 func (d mockDB) ReadProjectEntry(ctx context.Context, project string) (db.ProjectEntry, error) {
+	if project == projectDoesNotExist {
+		return db.ProjectEntry{}, upper.ErrNoMoreRows
+	}
+
+	if project == "projectreaderror" {
+		return db.ProjectEntry{}, errors.New("error reading DB")
+	}
+
 	return db.ProjectEntry{}, nil
 }
 
@@ -57,6 +102,14 @@ func (d mockDB) DeleteProjectEntry(ctx context.Context, project string) error {
 	}
 
 	return nil
+}
+
+func (d mockDB) DeleteTokenEntry(ctx context.Context, token string) error {
+	return nil
+}
+
+func (d mockDB) ReadTokenEntry(ctx context.Context, token string) (db.TokenEntry, error) {
+	return db.TokenEntry{}, nil
 }
 
 type mockGitClient struct{}
@@ -103,12 +156,24 @@ func newMockProvider(a credentials.Authorization, env env.Vars, h http.Header, f
 
 type mockCredentialsProvider struct{}
 
+func (m mockCredentialsProvider) DeleteProjectToken(projectName, tokenID string) error {
+	return nil
+}
+
+func (m mockCredentialsProvider) GetProjectToken(projectName, tokenID string) (types.ProjectToken, error) {
+	return types.ProjectToken{}, nil
+}
+
 func (m mockCredentialsProvider) GetToken() (string, error) {
 	return testPassword, nil
 }
 
-func (m mockCredentialsProvider) CreateProject(name string) (string, string, error) {
-	return "", "", nil
+func (m mockCredentialsProvider) CreateProject(name string) (string, string, string, error) {
+	return "role-id", "secret", "secret-id-accessor", nil
+}
+
+func (m mockCredentialsProvider) CreateToken(name string) (string, string, string, error) {
+	return "role-id", "secret", "secret-id-accessor", nil
 }
 
 func (m mockCredentialsProvider) DeleteProject(name string) error {
@@ -119,7 +184,7 @@ func (m mockCredentialsProvider) DeleteProject(name string) error {
 }
 
 func (m mockCredentialsProvider) GetProject(proj string) (responses.GetProject, error) {
-	if proj == "projectdoesnotexist" {
+	if proj == projectDoesNotExist {
 		return responses.GetProject{}, credentials.ErrNotFound
 	}
 	return responses.GetProject{Name: "project1"}, nil
@@ -167,6 +232,11 @@ func (m mockCredentialsProvider) ProjectExists(name string) (bool, error) {
 		"undeletableprojecttargets",
 		"undeletableproject",
 		"somedeletedberror",
+		"tokendberror",
+		"projectnotokens",
+		"projectreaderror",
+		"projectlisttokenserror",
+		"projectlisttokenslimit",
 	}
 	for _, existingProjects := range existingProjects {
 		if name == existingProjects {
@@ -196,6 +266,8 @@ type test struct {
 	authHeader string
 	url        string
 	method     string
+	dbMock     *th.DBClientMock
+	cpMock     *th.CredsProviderMock
 }
 
 func TestCreateProject(t *testing.T) {
@@ -241,6 +313,74 @@ func TestCreateProject(t *testing.T) {
 			want:       http.StatusInternalServerError,
 			authHeader: adminAuthHeader,
 			url:        "/projects",
+			method:     "POST",
+		},
+		{
+			name:       "project fails to create token entry",
+			req:        loadJSON(t, "TestCreateProject/project_fails_to_create_token_entry.json"),
+			want:       http.StatusInternalServerError,
+			authHeader: adminAuthHeader,
+			url:        "/projects",
+			method:     "POST",
+		},
+	}
+	runTests(t, tests)
+}
+
+func TestCreateToken(t *testing.T) {
+	tests := []test{
+
+		{
+			name:       "can create token",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusOK,
+			respFile:   "TestCreateToken/can_create_token_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "project does not exist",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusNotFound,
+			respFile:   "TestCreateToken/project_does_not_exist.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project1234/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "fails to create token when not admin",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusUnauthorized,
+			respFile:   "TestCreateToken/fails_to_create_token_when_not_admin_response.json",
+			authHeader: userAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "token fails to create db entry",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusInternalServerError,
+			authHeader: adminAuthHeader,
+			url:        "/projects/tokendberror/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "allowed tokens limit reached",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusInternalServerError,
+			respFile:   "TestCreateToken/token_limit_reached_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectlisttokenslimit/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "error listing tokens",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusInternalServerError,
+			respFile:   "TestCreateToken/error_listing_tokens_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectlisttokenserror/tokens",
 			method:     "POST",
 		},
 	}
@@ -684,6 +824,215 @@ func TestListWorkflows(t *testing.T) {
 	runTests(t, tests)
 }
 
+func TestDeleteToken(t *testing.T) {
+	tests := []test{
+		{
+			name:       "fails to delete tokens when not admin",
+			want:       http.StatusUnauthorized,
+			respFile:   "TestDeleteToken/fails_to_delete_token_when_not_admin_response.json",
+			authHeader: userAuthHeader,
+			url:        "/projects/project/tokens/existingtoken",
+			method:     "DELETE",
+		},
+		{
+			name:       "can delete token",
+			want:       http.StatusOK,
+			respFile:   "TestDeleteToken/can_delete_token_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/existingtoken",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{ID: "1234"}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+				DeleteProjectTokenFunc: func(p, t string) error {
+					return nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				DeleteTokenEntryFunc: func(ctx context.Context, token string) error {
+					return nil
+				},
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ReadTokenEntryFunc: func(ctx context.Context, token string) (db.TokenEntry, error) {
+					return db.TokenEntry{ProjectID: "project1", TokenID: "1234", CreatedAt: "2022-06-21T14:42:50.182037-07:00"}, nil
+				},
+			},
+		},
+		{
+			name:       "project does not exist",
+			want:       http.StatusNotFound,
+			respFile:   "TestDeleteToken/project_does_not_exist_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectdoesnotexist/tokens/tokendoesnotexist",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return false, nil
+				},
+			},
+		},
+		{
+			name:       "token does not exist",
+			want:       http.StatusNotFound,
+			respFile:   "TestDeleteToken/token_does_not_exist_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/tokendoesnotexist",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+			},
+		},
+		{
+			name:       "token delete error",
+			want:       http.StatusInternalServerError,
+			respFile:   "TestDeleteToken/token_delete_error_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/deletetokenerror",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				DeleteProjectTokenFunc: func(s1, s2 string) error {
+					return errors.New("error deleting token from Vault")
+				},
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{ID: "1234"}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				DeleteTokenEntryFunc: func(ctx context.Context, token string) error {
+					return errors.New("error deleting entry from DB")
+				},
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ReadTokenEntryFunc: func(ctx context.Context, token string) (db.TokenEntry, error) {
+					return db.TokenEntry{ProjectID: "project1", TokenID: "1234", CreatedAt: "2022-06-21T14:42:50.182037-07:00"}, nil
+				},
+			},
+		},
+	}
+	runTestsV2(t, tests)
+}
+
+func TestListTokens(t *testing.T) {
+	tests := []test{
+		{
+			name:       "fails to list tokens when not admin",
+			want:       http.StatusUnauthorized,
+			respFile:   "TestListTokens/fails_to_list_tokens_when_not_admin_response.json",
+			authHeader: userAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "GET",
+		},
+		{
+			name:       "can list tokens",
+			want:       http.StatusOK,
+			respFile:   "TestListTokens/can_list_tokens_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, p string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1", Repository: "repo"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{
+						{
+							CreatedAt: "2022-06-21T14:56:10.341066-07:00",
+							TokenID:   "ghi789",
+						},
+						{
+							CreatedAt: "2022-06-21T14:43:16.172896-07:00",
+							TokenID:   "def456",
+						},
+						{
+							CreatedAt: "2022-06-21T14:42:50.182037-07:00",
+							TokenID:   "abc123",
+						},
+					}, nil
+				},
+			},
+		},
+		{
+			name:       "project not found",
+			want:       http.StatusNotFound,
+			respFile:   "TestListTokens/project_not_found_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectdoesnotexist/tokens",
+			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, p string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{}, upper.ErrNoMoreRows
+				},
+			},
+		},
+		{
+			name:       "no tokens",
+			want:       http.StatusOK,
+			respFile:   "TestListTokens/no_tokens_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectnotokens/tokens",
+			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, p string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "abc123", Repository: "repo"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{}, nil
+				},
+			},
+		},
+		{
+			name:       "project read error",
+			want:       http.StatusInternalServerError,
+			respFile:   "TestListTokens/project_read_error_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectreaderror/tokens",
+			method:     "GET",
+			cpMock: &th.CredsProviderMock{
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return false, errors.New("error retrieving project")
+				},
+			},
+		},
+		{
+			name:       "list tokens read error",
+			want:       http.StatusInternalServerError,
+			respFile:   "TestListTokens/list_tokens_error_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectlisttokenserror/tokens",
+			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{}, errors.New("error from DB")
+				},
+			},
+		},
+	}
+	runTestsV2(t, tests)
+}
+
 func TestHealthCheck(t *testing.T) {
 	tests := []struct {
 		name                  string
@@ -820,7 +1169,89 @@ func runTests(t *testing.T, tests []test) {
 
 				defer resp.Body.Close()
 
-				assert.JSONEq(t, string(wantBody), string(body))
+				bodyStr := string(body)
+				wantBodyStr := string(wantBody)
+
+				if bodyStr == "" && wantBodyStr == "" {
+					assert.Equal(t, wantBodyStr, bodyStr)
+				} else {
+					assert.JSONEq(t, wantBodyStr, bodyStr)
+				}
+			}
+		})
+	}
+}
+
+func runTestsV2(t *testing.T, tests []test) {
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			config, err := loadConfig(testConfigPath)
+			if err != nil {
+				panic(fmt.Sprintf("Unable to load config %s", err))
+			}
+
+			h := handler{
+				logger:                 log.NewNopLogger(),
+				newCredentialsProvider: newMockProvider,
+				argo:                   mockWorkflowSvc{},
+				argoCtx:                context.Background(),
+				config:                 config,
+				gitClient:              newMockGitClient(),
+				env: env.Vars{
+					AdminSecret: testPassword,
+				},
+				dbClient: newMockDB(),
+			}
+
+			if tt.dbMock != nil {
+				h.dbClient = tt.dbMock
+			}
+
+			if tt.cpMock != nil {
+				defaultCPFunc := func(a credentials.Authorization, env env.Vars, h http.Header, f credentials.VaultConfigFn, fn credentials.VaultSvcFn) (credentials.Provider, error) {
+					// TODO: probably best to create a base/default CP mock struct here
+					return tt.cpMock, nil
+				}
+
+				h.newCredentialsProvider = defaultCPFunc
+			}
+
+			resp := executeRequestWithHandler(h, tt.method, tt.url, serialize(tt.req), tt.authHeader)
+			if resp.StatusCode != tt.want {
+				t.Errorf("Unexpected status code %d", resp.StatusCode)
+			}
+
+			if tt.body != "" {
+				bodyBytes, err := ioutil.ReadAll(resp.Body)
+				defer resp.Body.Close()
+				if err != nil {
+					t.Errorf("Error loading body")
+				}
+				if tt.body != string(bodyBytes) {
+					t.Errorf("Unexpected body '%s', expected '%s'", bodyBytes, tt.body)
+				}
+			}
+
+			if tt.respFile != "" {
+				wantBody, err := loadFileBytes(tt.respFile)
+				if err != nil {
+					t.Fatalf("unable to read response file '%s', err: '%s'", tt.respFile, err)
+				}
+
+				body, err := io.ReadAll(resp.Body)
+				assert.Nil(t, err)
+
+				defer resp.Body.Close()
+
+				bodyStr := string(body)
+				wantBodyStr := string(wantBody)
+
+				// don't use assert.JSONEq for empty strings
+				if bodyStr == "" && wantBodyStr == "" {
+					assert.Equal(t, wantBodyStr, bodyStr)
+				} else {
+					assert.JSONEq(t, wantBodyStr, bodyStr)
+				}
 			}
 		})
 	}
@@ -855,9 +1286,19 @@ func executeRequest(method string, url string, body *bytes.Buffer, authHeader st
 	return w.Result()
 }
 
+func executeRequestWithHandler(h handler, method string, url string, body *bytes.Buffer, authHeader string) *http.Response {
+	var router = setupRouter(h)
+	req, _ := http.NewRequest(method, url, serialize(body))
+
+	req.Header.Add("Authorization", authHeader)
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, req)
+	return w.Result()
+}
+
 // loadFileBytes returns the contents of a file in the 'testdata' directory.
 func loadFileBytes(filename string) ([]byte, error) {
-	file := filepath.Join("testdata", filename)
+	file := filepath.Join("test/testdata", filename)
 	return os.ReadFile(file)
 }