+++ service/internal/credentials/vault.go
@@ -1,3 +1,5 @@
+//go:generate moq -out ../../test/testhelpers/credsProviderMock.go -pkg testhelpers . Provider:CredsProviderMock
+
 package credentials
 
 import (
@@ -20,14 +22,17 @@ const (
 
 // Provider defines the interface required by providers.
 type Provider interface {
-	CreateProject(string) (string, string, error)
+	CreateProject(string) (string, string, string, error)
 	CreateTarget(string, types.Target) error
+	CreateToken(string) (string, string, string, error)
 	UpdateTarget(string, types.Target) error
 	DeleteProject(string) error
 	DeleteTarget(string, string) error
 	GetProject(string) (responses.GetProject, error)
 	GetTarget(string, string) (types.Target, error)
 	GetToken() (string, error)
+	DeleteProjectToken(string, string) error
+	GetProjectToken(string, string) (types.ProjectToken, error)
 	ListTargets(string) ([]string, error)
 	ProjectExists(string) (bool, error)
 	TargetExists(string, string) (bool, error)
@@ -186,32 +191,40 @@ func genProjectAppRole(name string) string {
 	return fmt.Sprintf("%s/%s-%s", vaultAppRolePrefix, vaultProjectPrefix, name)
 }
 
-func (v VaultProvider) CreateProject(name string) (string, string, error) {
+func (v VaultProvider) CreateToken(name string) (string, string, string, error) {
 	if !v.isAdmin() {
-		return "", "", errors.New("admin credentials must be used to create project")
+		return "", "", "", errors.New("admin credentials must be used to create token")
+	}
+
+	secretID, secretAccessor, err := v.generateSecrets(name)
+	if err != nil {
+		return "", "", "", err
+	}
+
+	roleID, err := v.readRoleID(name)
+	if err != nil {
+		return "", "", "", err
+	}
+
+	return roleID, secretID, secretAccessor, nil
+}
+
+func (v VaultProvider) CreateProject(name string) (string, string, string, error) {
+	if !v.isAdmin() {
+		return "", "", "", errors.New("admin credentials must be used to create project")
 	}
 
 	policy := defaultVaultReadonlyPolicyAWS(name)
 	err := v.createPolicyState(name, policy)
 	if err != nil {
-		return "", "", err
+		return "", "", "", err
 	}
 
 	if err := v.writeProjectState(name); err != nil {
-		return "", "", err
+		return "", "", "", err
 	}
 
-	secretID, err := v.readSecretID(name)
-	if err != nil {
-		return "", "", err
-	}
-
-	roleID, err := v.readRoleID(name)
-	if err != nil {
-		return "", "", err
-	}
-
-	return roleID, secretID, nil
+	return v.CreateToken(name)
 }
 
 // CreateTarget creates a target for the project.
@@ -336,6 +349,50 @@ func (v VaultProvider) GetTarget(projectName, targetName string) (types.Target, 
 	}, nil
 }
 
+func (v VaultProvider) DeleteProjectToken(projectName, tokenID string) error {
+	if !v.isAdmin() {
+		return errors.New("admin credentials must be used to delete tokens")
+	}
+
+	data := map[string]interface{}{
+		"secret_id_accessor": tokenID,
+	}
+
+	path := fmt.Sprintf("%s/secret-id-accessor/destroy", genProjectAppRole(projectName))
+	_, err := v.vaultLogicalSvc.Write(path, data)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (v VaultProvider) GetProjectToken(projectName, tokenID string) (types.ProjectToken, error) {
+	token := types.ProjectToken{}
+
+	if !v.isAdmin() {
+		return token, errors.New("admin credentials must be used to delete tokens")
+	}
+
+	data := map[string]interface{}{
+		"secret_id_accessor": tokenID,
+	}
+
+	path := fmt.Sprintf("%s/secret-id-accessor/lookup", genProjectAppRole(projectName))
+	projectToken, err := v.vaultLogicalSvc.Write(path, data)
+	if err != nil {
+		return token, fmt.Errorf("vault get secret ID accessor error: %w", err)
+	}
+
+	if projectToken == nil {
+		return token, nil
+	}
+
+	return types.ProjectToken{
+		ID: projectToken.Data["secret_id_accessor"].(string),
+	}, nil
+}
+
 func (v VaultProvider) GetToken() (string, error) {
 	if v.isAdmin() {
 		return "", errors.New("admin credentials cannot be used to get tokens")
@@ -406,15 +463,15 @@ func (v VaultProvider) readRoleID(appRoleName string) (string, error) {
 	return secret.Data["role_id"].(string), nil
 }
 
-func (v VaultProvider) readSecretID(appRoleName string) (string, error) {
+func (v VaultProvider) generateSecrets(appRoleName string) (string, string, error) {
 	options := map[string]interface{}{
 		"force": true,
 	}
 	secret, err := v.vaultLogicalSvc.Write(fmt.Sprintf("%s/secret-id", genProjectAppRole(appRoleName)), options)
 	if err != nil {
-		return "", err
+		return "", "", err
 	}
-	return secret.Data["secret_id"].(string), nil
+	return secret.Data["secret_id"].(string), secret.Data["secret_id_accessor"].(string), nil
 }
 
 func (v VaultProvider) TargetExists(projectName, targetName string) (bool, error) {