+++ CHANGELOG.md
@@ -4,11 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2022-07-08
+### Added
+* Added schema updates to create tokens table
+* Added golang-migrate to manage schema
+* Delete project token
+* List project tokens
+* Create project token
+
+## [0.13.3] - 2022-06-16
+### Changed
+* Added github repository from db to get projects response
+
 ## [0.13.2] - 2022-06-09
 ### Changed
 * Fixed quickstart issues with standing up database
 * Added log statement to clarify issues with project creation
-
 
 ## [0.13.1] - 2022-05-24
 ### Changed